Fixed error where S&T are set falsely (to E1 and E2 instead) for E6POS.
This bug happens because the super class of KRLE6Pos, which is KRLPos, uses
the getNodes() array and takes the seventh and eighth element of that list EVERY TIME cause it assumes that S and T are stored there. But for E6POS there are the E1-6 elements which start at nodes[6] and go through nodes[11]. For E6POS objects the S and T are therefore at the nodes[12] and nodes[13] position.

Fixed that bug by specifying the wanted key directly instead of the relativity that getNodes() creates.